@@ -1,13 +1,16 @@
 package snappy
 
 import (
-	"github.com/stretchr/testify/assert"
-	tmjson "github.com/tendermint/tendermint/libs/json"
-	tendermint "github.com/tendermint/tendermint/types"
-	db "github.com/tendermint/tm-db"
-	"io/ioutil"
+	"io"
 	"os"
 	"testing"
+
+	cbftjson "github.com/cometbft/cometbft/libs/json"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/cometbft/cometbft/types"
+
+	db "github.com/cometbft/cometbft-db"
 )
 
 func TestSnappyDB(t *testing.T) {
@@ -81,10 +84,10 @@ func TestSnappyDBCompat(t *testing.T) {
 }
 
 func indexSampleTx(mdb db.DB, key []byte) {
-	block := &tendermint.Block{}
+	block := &types.Block{}
 	blockFile, _ := os.Open("../../indexer/fixtures/block_4814775.json")
-	blockJSON, _ := ioutil.ReadAll(blockFile)
-	if err := tmjson.Unmarshal(blockJSON, block); err != nil {
+	blockJSON, _ := io.ReadAll(blockFile)
+	if err := cbftjson.Unmarshal(blockJSON, block); err != nil {
 		panic(err)
 	}
 
