@@ -3,11 +3,13 @@ package hld
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"sync"
 
+	"github.com/cosmos/cosmos-sdk/store/types"
 	"github.com/terra-money/mantlemint/lib"
 
-	tmdb "github.com/tendermint/tm-db"
+	dbm "github.com/cometbft/cometbft-db"
 )
 
 const (
@@ -29,6 +31,7 @@ var (
 )
 
 var _ HLD = (*HeightLimitedDB)(nil)
+// var _ types.CommitKVStore = (*HeightLimitedDB)(nil)
 
 type HeightLimitedDB struct {
 	odb         HeightLimitEnabledDB
@@ -39,6 +42,12 @@ type HeightLimitedDB struct {
 	config      *HeightLimitedDBConfig
 
 	// writeBatch HeightLimitEnabledBatch
+}
+
+// Compact implements HLD.
+func (hld *HeightLimitedDB) Compact(start []byte, end []byte) error {
+	// TODO: do we need to implement this?
+	return nil
 }
 
 type HeightLimitedDBConfig struct {
@@ -143,7 +152,7 @@ func (hld *HeightLimitedDB) Delete(key []byte) error {
 
 // DeleteSync deletes the key, and flushes the delete to storage before returning.
 func (hld *HeightLimitedDB) DeleteSync(key []byte) error {
-	return hld.Delete(key)
+	return hld.odb.DeleteSync(hld.writeHeight, key)
 }
 
 // Iterator returns an iterator over a domain of keys, in ascending order. The caller must call
@@ -151,7 +160,7 @@ func (hld *HeightLimitedDB) DeleteSync(key []byte) error {
 // from the first key, and a nil end iterates to the last key (inclusive).
 // CONTRACT: No writes may happen within a domain while an iterator exists over it.
 // CONTRACT: start, end readonly []byte
-func (hld *HeightLimitedDB) Iterator(start, end []byte) (tmdb.Iterator, error) {
+func (hld *HeightLimitedDB) Iterator(start, end []byte) (dbm.Iterator, error) {
 	return hld.odb.Iterator(hld.GetCurrentReadHeight(), start, end)
 }
 
@@ -160,7 +169,7 @@ func (hld *HeightLimitedDB) Iterator(start, end []byte) (tmdb.Iterator, error) {
 // iterates from the last key (inclusive), and a nil start iterates to the first key (inclusive).
 // CONTRACT: No writes may happen within a domain while an iterator exists over it.
 // CONTRACT: start, end readonly []byte
-func (hld *HeightLimitedDB) ReverseIterator(start, end []byte) (tmdb.Iterator, error) {
+func (hld *HeightLimitedDB) ReverseIterator(start, end []byte) (dbm.Iterator, error) {
 	return hld.odb.ReverseIterator(hld.GetCurrentReadHeight(), start, end)
 }
 
@@ -170,7 +179,7 @@ func (hld *HeightLimitedDB) Close() error {
 }
 
 // NewBatch creates a batch for atomic updates. The caller must call Batch.Close.
-func (hld *HeightLimitedDB) NewBatch() tmdb.Batch {
+func (hld *HeightLimitedDB) NewBatch() dbm.Batch {
 	// if hld.writeBatch != nil {
 	// 	// TODO: fix me
 	// 	return hld.writeBatch
@@ -230,4 +239,28 @@ func (hld *HeightLimitedDB) Debug(debugType int, key []byte, value []byte) {
 
 	fmt.Printf("<%s @ %d> %s", debugPrefix, actualKeyHeight, keyFamily)
 	fmt.Printf("\n")
+}
+
+func (hld *HeightLimitedDB) CacheWrap() types.CacheWrap {
+	panic("unimplemented")
+}
+
+func (hld *HeightLimitedDB) CacheWrapWithTrace(w io.Writer, tc types.TraceContext) types.CacheWrap {
+	panic("unimplemented")
+}
+
+// Commit() CommitID
+// LastCommitID() CommitID
+// SetCommitting()
+// UnsetCommitting()
+
+// SetPruning(pruningtypes.PruningOptions)
+// GetPruning() pruningtypes.PruningOptions
+
+func (hld *HeightLimitedDB) Commit() types.CommitID {
+	panic("unimplemented")
+}
+
+func (hld *HeightLimitedDB) LastCommitID() types.CommitID {
+	panic("unimplemented")
 }
