@@ -3,22 +3,22 @@ package block
 import (
 	"fmt"
 
-	tmjson "github.com/tendermint/tendermint/libs/json"
-	tm "github.com/tendermint/tendermint/types"
-	terra "github.com/terra-money/core/v2/app"
-	"github.com/terra-money/mantlemint/db/safe_batch"
-	"github.com/terra-money/mantlemint/indexer"
-	"github.com/terra-money/mantlemint/mantlemint"
+	cbftjson "github.com/cometbft/cometbft/libs/json"
+	"github.com/cometbft/cometbft/types"
+	"github.com/osmosis-labs/mantlemint/db/safe_batch"
+	"github.com/osmosis-labs/mantlemint/indexer"
+	"github.com/osmosis-labs/mantlemint/mantlemint"
+	"github.com/osmosis-labs/osmosis/v25/app"
 )
 
-var IndexBlock = indexer.CreateIndexer(func(indexerDB safe_batch.SafeBatchDB, block *tm.Block, blockID *tm.BlockID, _ *mantlemint.EventCollector, _ *terra.TerraApp) error {
+var IndexBlock = indexer.CreateIndexer(func(indexerDB safe_batch.SafeBatchDB, block *types.Block, blockID *types.BlockID, _ *mantlemint.EventCollector, _ *app.OsmosisApp) error {
 	defer fmt.Printf("[indexer/block] indexing done for height %d\n", block.Height)
 	record := BlockRecord{
 		Block:   block,
 		BlockID: blockID,
 	}
 
-	recordJSON, recordErr := tmjson.Marshal(record)
+	recordJSON, recordErr := cbftjson.Marshal(record)
 	if recordErr != nil {
 		return recordErr
 	}
