@@ -3,13 +3,14 @@ package block
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/pkg/errors"
 	"net/http"
 	"strconv"
 
+	"github.com/osmosis-labs/mantlemint/indexer"
+	"github.com/pkg/errors"
+
+	cbftdb "github.com/cometbft/cometbft-db"
 	"github.com/gorilla/mux"
-	tmdb "github.com/tendermint/tm-db"
-	"github.com/terra-money/mantlemint/indexer"
 )
 
 var (
@@ -21,7 +22,7 @@ var (
 	ErrorBlockNotFound = func(height string) string { return fmt.Sprintf("block %s not found... yet.", height) }
 )
 
-func blockByHeightHandler(indexerDB tmdb.DB, height string) (json.RawMessage, error) {
+func blockByHeightHandler(indexerDB cbftdb.DB, height string) (json.RawMessage, error) {
 	heightInInt, err := strconv.Atoi(height)
 	if err != nil {
 		return nil, errors.New(ErrorInvalidHeight(height))
@@ -29,7 +30,7 @@ func blockByHeightHandler(indexerDB tmdb.DB, height string) (json.RawMessage, er
 	return indexerDB.Get(getKey(uint64(heightInInt)))
 }
 
-var RegisterRESTRoute = indexer.CreateRESTRoute(func(router *mux.Router, indexerDB tmdb.DB) {
+var RegisterRESTRoute = indexer.CreateRESTRoute(func(router *mux.Router, indexerDB cbftdb.DB) {
 	router.HandleFunc(EndpointGETBlocksHeight, func(writer http.ResponseWriter, request *http.Request) {
 		vars := mux.Vars(request)
 		height, ok := vars["height"]
