@@ -1,8 +1,8 @@
 package block
 
 import (
-	tm "github.com/tendermint/tendermint/types"
-	"github.com/terra-money/mantlemint/lib"
+	cbfttypes "github.com/cometbft/cometbft/types"
+	"github.com/osmosis-labs/mantlemint/lib"
 )
 
 var prefix = []byte("block/height:")
@@ -11,6 +11,6 @@ var getKey = func(height uint64) []byte {
 }
 
 type BlockRecord struct {
-	BlockID *tm.BlockID `json:"block_id"`
-	Block   *tm.Block   `json:"block"`
+	BlockID *cbfttypes.BlockID `json:"block_id"`
+	Block   *cbfttypes.Block   `json:"block"`
 }
