@@ -8,9 +8,9 @@ import (
 
 	"github.com/pkg/errors"
 
+	cbftdb "github.com/cometbft/cometbft-db"
 	"github.com/gorilla/mux"
-	tmdb "github.com/tendermint/tm-db"
-	"github.com/terra-money/mantlemint/indexer"
+	"github.com/osmosis-labs/mantlemint/indexer"
 )
 
 var (
@@ -22,7 +22,7 @@ var (
 	ErrorRichlistNotFound = func(height string) string { return fmt.Sprintf("richlist at %s not found... yet.", height) }
 )
 
-func richlistByHeightHandler(indexerDB tmdb.DB, height string) (json.RawMessage, error) {
+func richlistByHeightHandler(indexerDB cbftdb.DB, height string) (json.RawMessage, error) {
 	heightInInt, err := strconv.Atoi(height)
 	if err != nil {
 		return nil, errors.New(ErrorInvalidHeight(height))
@@ -30,7 +30,7 @@ func richlistByHeightHandler(indexerDB tmdb.DB, height string) (json.RawMessage,
 	return indexerDB.Get(getDefaultKey(uint64(heightInInt)))
 }
 
-var RegisterRESTRoute = indexer.CreateRESTRoute(func(router *mux.Router, indexerDB tmdb.DB) {
+var RegisterRESTRoute = indexer.CreateRESTRoute(func(router *mux.Router, indexerDB cbftdb.DB) {
 	router.HandleFunc(EndpointGETBlocksHeight, func(writer http.ResponseWriter, request *http.Request) {
 		vars := mux.Vars(request)
 		height, ok := vars["height"]
