@@ -5,12 +5,13 @@ import (
 	"fmt"
 	"sync"
 
+	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/google/btree"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-	terra "github.com/terra-money/core/v2/app"
-	"github.com/terra-money/mantlemint/lib"
+	"github.com/osmosis-labs/mantlemint/lib"
+
+	"github.com/osmosis-labs/osmosis/v25/app"
 )
 
 // keys for storage
@@ -182,7 +183,7 @@ func (list *Richlist) Extract(height uint64, len int, threshold *sdk.Coin) (extr
 	return
 }
 
-func (list *Richlist) Apply(changes map[string]sdk.Int, app *terra.TerraApp, height uint64, denom string) (err error) {
+func (list *Richlist) Apply(changes map[string]sdk.Int, app *app.OsmosisApp, height uint64, denom string) (err error) {
 	ctx := app.NewContext(true, tmproto.Header{})
 
 	for address, amount := range changes {
