@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"time"
 
-	abci "github.com/tendermint/tendermint/abci/types"
-	"github.com/terra-money/mantlemint/lib"
+	abci "github.com/cometbft/cometbft/abci/types"
+	"github.com/osmosis-labs/mantlemint/lib"
 )
 
 type TxRecord struct {
