@@ -11,20 +11,20 @@ import (
 	"github.com/cosmos/cosmos-sdk/server/api"
 	"github.com/cosmos/cosmos-sdk/server/config"
 
+	tmlog "github.com/cometbft/cometbft/libs/log"
+	rpcclient "github.com/cometbft/cometbft/rpc/client"
 	"github.com/cosmos/cosmos-sdk/server/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/gorilla/mux"
+	mconfig "github.com/osmosis-labs/mantlemint/config"
+	"github.com/osmosis-labs/mantlemint/export"
+	osmosisapp "github.com/osmosis-labs/osmosis/v25/app"
+	"github.com/osmosis-labs/osmosis/v25/app/params"
 	"github.com/spf13/viper"
-	tmlog "github.com/tendermint/tendermint/libs/log"
-	rpcclient "github.com/tendermint/tendermint/rpc/client"
-	terra "github.com/terra-money/core/v2/app"
-	"github.com/terra-money/core/v2/app/params"
-	mconfig "github.com/terra-money/mantlemint/config"
-	"github.com/terra-money/mantlemint/export"
 )
 
 func StartRPC(
-	app *terra.TerraApp,
+	app *osmosisapp.OsmosisApp,
 	rpcclient rpcclient.Client,
 	chainId string,
 	codec params.EncodingConfig,
@@ -36,7 +36,7 @@ func StartRPC(
 	vp := viper.GetViper()
 	cfg, _ := config.GetConfig(vp)
 
-	// create terra client; register all codecs
+	// create osmosis client; register all codecs
 	context := client.
 		Context{}.
 		WithClient(rpcclient).
@@ -45,7 +45,7 @@ func StartRPC(
 		WithTxConfig(codec.TxConfig).
 		WithAccountRetriever(authtypes.AccountRetriever{}).
 		WithLegacyAmino(codec.Amino).
-		WithHomeDir(terra.DefaultNodeHome).
+		WithHomeDir(osmosisapp.DefaultNodeHome).
 		WithChainID(chainId)
 
 	// create backends for response cache
